perf: don't load reports on home page
